refactor: restructure the dump load code
